+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Fixed `FormattedInputCurrency` invalid input is changed to 0 on blur
 
 ## 1.2.1
 * Upgraded `@opuscapita/format-utils` version